last commit on 25 mordad 1401 .chat page completed

+++ lib/Models/Home/pricey_courses_model.dart
@@ -36,7 +36,7 @@ class PriceyCoursesModel {
   String? views;
   List<int>? update;
   int? reviews;
-  double? reviewsRating;
+  int? reviewsRating;
 
   RxBool visible;
